Compiled src/OpenFOAM/global.

+++ src/OpenFOAM/global/profiling/profiling.C
@@ -96,8 +96,8 @@ Foam::profilingInformation* Foam::profiling::endTimer()
 {
     Information *info = stack_.remove();
     clockValue clockval = times_.remove();
-
-    info->update(clockval.elapsed());   // Update elapsed time
+    // CoDiPack4OpenFOAM
+    info->update(static_cast<double>(clockval.elapsed()));   // Update elapsed time
     info->setActive(false);             // Mark as off stack
 
     return info;
